Initial work to support adding using config files from other frameworks
