+++ GenericSynthesisPatcher/Games/Fallout4/Json/Action/ContainerItemsData.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel;
 
 using GenericSynthesisPatcher.Games.Universal.Json.Action;
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Mutagen.Bethesda.Fallout4;
 using Mutagen.Bethesda.Plugins.Records;
