+++ GenericSynthesisPatcher/Games/Skyrim/Json/Action/LeveledNpcData.cs
@@ -1,4 +1,4 @@
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Mutagen.Bethesda.Plugins.Records;
 using Mutagen.Bethesda.Skyrim;
