+++ GenericSynthesisPatcher/Games/Skyrim/Json/Action/RelationsData.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel;
 
 using GenericSynthesisPatcher.Games.Universal.Json.Action;
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Mutagen.Bethesda.Plugins.Records;
 using Mutagen.Bethesda.Skyrim;
