+++ GenericSynthesisPatcher/Games/Universal/Action/ConvertibleAction.cs
@@ -2,8 +2,8 @@ using System.Diagnostics.CodeAnalysis;
 
 using Common;
 
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
 using GenericSynthesisPatcher.Helpers;
+using GenericSynthesisPatcher.Rules.Operations;
 
 namespace GenericSynthesisPatcher.Games.Universal.Action
 {
