+++ GenericSynthesisPatcher/Games/Universal/Action/FlagsAction.cs
@@ -4,9 +4,9 @@ using Common;
 
 using EnumsNET;
 
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
 using GenericSynthesisPatcher.Helpers;
 using GenericSynthesisPatcher.Helpers.Graph;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Microsoft.Extensions.Logging;
