+++ GenericSynthesisPatcher/Games/Universal/Json/Action/FormLinksWithDataActionDataBase.cs
@@ -1,4 +1,4 @@
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Mutagen.Bethesda.Plugins.Records;
