+++ GenericSynthesisPatcher/Games/Universal/Json/Converters/OperationsConverter.cs
@@ -1,4 +1,4 @@
-using GenericSynthesisPatcher.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Microsoft.Extensions.Logging;
