--- GenericSynthesisPatcher/Games/Universal/Json/Data/GSPRule.cs
@@ -1,461 +1,0 @@
-using System.Diagnostics.CodeAnalysis;
-
-using Common;
-
-using GenericSynthesisPatcher.Games.Universal.Action;
-using GenericSynthesisPatcher.Games.Universal.Json.Converters;
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
-using GenericSynthesisPatcher.Helpers;
-
-using Microsoft.Extensions.Logging;
-
-using Mutagen.Bethesda.Plugins;
-
-using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
-
-using Noggog;
-
-namespace GenericSynthesisPatcher.Games.Universal.Json.Data
-{
-    public class GSPRule : GSPBase
-    {
-        private const int ClassLogCode = 0x1C;
-        private List<ListOperation>? editorIDs;
-        private List<FormKeyListOperation>? formIDs;
-        private int HashCode;
-
-        /// <summary>
-        ///     Details for DeepCopIn Action if required by this rule.
-        /// </summary>
-        [JsonProperty(PropertyName = "Copy", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<GSPDeepCopyIn>))]
-        public List<GSPDeepCopyIn> DeepCopyIn { get; set; } = [];
-
-        /// <summary>
-        ///     List of EditorIDs to match against that will be included or excluded. If not set,
-        ///     not filter any records out by EditorID.
-        /// </summary>
-        [JsonProperty(PropertyName = "EditorID", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<ListOperation>))]
-        public List<ListOperation>? EditorID
-        {
-            get => editorIDs;
-            set
-            {
-                if (!value.SafeAny())
-                    return;
-
-                editorIDs = value;
-            }
-        }
-
-        [JsonProperty(PropertyName = "-EditorID", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<ListOperation>))]
-        public List<ListOperation>? EditorIDDel
-        {
-            set
-            {
-                if (!value.SafeAny())
-                    return;
-
-                editorIDs = [];
-                foreach (var v in value)
-                    editorIDs.Add(v.Inverse());
-            }
-        }
-
-        [JsonProperty(PropertyName = "!EditorID", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<ListOperation>))]
-        public List<ListOperation>? EditorIDNot { set => EditorIDDel = value; }
-
-        /// <summary>
-        ///     Add Fill action(s) to this rule. This is only used for adding to joint Fill/Forward store.
-        /// </summary>
-        [JsonProperty(PropertyName = "Fill", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(DictionaryConverter<FilterOperation, JToken>))]
-        public Dictionary<FilterOperation, JToken> Fill { get; set; } = [];
-
-        /// <summary>
-        ///     List of FormIDs to match against that will be included or excluded. If not set, not
-        ///     filter any records out by FormID.
-        /// </summary>
-        [JsonProperty(PropertyName = "FormID", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<FormKeyListOperation>))]
-        public List<FormKeyListOperation>? FormID
-        {
-            get => formIDs;
-            set
-            {
-                if (!value.SafeAny())
-                    return;
-
-                formIDs = value;
-            }
-        }
-
-        [JsonProperty(PropertyName = "-FormID", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<FormKeyListOperation>))]
-        public List<FormKeyListOperation>? FormIDDel
-        {
-            set
-            {
-                if (!value.SafeAny())
-                    return;
-
-                formIDs = [];
-                foreach (var v in value)
-                    formIDs.Add(v.Inverse());
-            }
-        }
-
-        [JsonProperty(PropertyName = "!FormID", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(SingleOrArrayConverter<FormKeyListOperation>))]
-        public List<FormKeyListOperation>? FormIDNot { set => FormIDDel = value; }
-
-        /// <summary>
-        ///     Add Forward action(s) to this rule. This is only used for adding to joint
-        ///     Fill/Forward store.
-        /// </summary>
-        [JsonProperty(PropertyName = "Forward", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(DictionaryConverter<FilterOperation, JToken>))]
-        public Dictionary<FilterOperation, JToken> Forward { get; set; } = [];
-
-        /// <summary>
-        ///     Changes JSON format for Forward from being the default "mod.esp" : [ "field1",
-        ///     "field2" ] to be "field": [ "Mod1.esp", "Mod2.esp" ]
-        /// </summary>
-        [JsonProperty(PropertyName = "ForwardIndexedByField", NullValueHandling = NullValueHandling.Ignore)]
-        [Obsolete("""ForwardIndexedByField is deprecated. Should use "ForwardFlags": ["IndexedByField"]""")]
-        public bool ForwardIndexedByField
-        {
-            get => ForwardOptions.HasFlag(ForwardOptions.IndexedByField);
-            set
-            {
-                Global.Logger.WriteLog(LogLevel.Warning, LogType.GeneralConfigFailure, """ForwardIndexedByField is deprecated. Should use "ForwardOptions": ["IndexedByField"]""", ClassLogCode);
-
-                ForwardOptions = ForwardOptions.SetFlag(ForwardOptions.IndexedByField, value);
-            }
-        }
-
-        /// <summary>
-        ///     ForwardType can change how Forwarding actions work.
-        /// </summary>
-        [JsonProperty(PropertyName = "ForwardOptions", NullValueHandling = NullValueHandling.Ignore)]
-        public ForwardOptions ForwardOptions { get; set; }
-
-        /// <summary>
-        ///     ForwardType can change how Forwarding actions work.
-        /// </summary>
-        [JsonProperty(PropertyName = "ForwardType", NullValueHandling = NullValueHandling.Ignore)]
-        [Obsolete("""ForwardType is deprecated. Should use "ForwardOptions": ["..."]""")]
-        public ForwardOptions ForwardType
-        {
-            get => ForwardOptions;
-            set
-            {
-                Global.Logger.WriteLog(LogLevel.Warning, LogType.GeneralConfigFailure, """ForwardType is deprecated. Should use "ForwardOptions": ["..."]""", ClassLogCode);
-
-                ForwardOptions |= value;
-            }
-        }
-
-        /// <summary>
-        ///     If rule belongs to a group, this will point to the group that it belongs to.
-        /// </summary>
-        [JsonIgnore]
-        public GSPGroup? Group { get; private set; } = null;
-
-        /// <summary>
-        ///     Add filters to this rule that can filter on supported properties that are not
-        ///     EditorID or FormID.
-        /// </summary>
-        [JsonProperty(PropertyName = "Matches", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(DictionaryConverter<FilterOperation, JToken>))]
-        public Dictionary<FilterOperation, JToken> Match { get; set; } = [];
-
-        /// <summary>
-        ///     List of merge actions to perform.
-        /// </summary>
-        [JsonProperty(PropertyName = "Merge", NullValueHandling = NullValueHandling.Ignore)]
-        [JsonConverter(typeof(DictionaryConverter<FilterOperation, List<ModKeyListOperation>>))]
-        public Dictionary<FilterOperation, List<ModKeyListOperation>?> Merge { get; set; } = [];
-
-        /// <summary>
-        ///     If set to true, will only apply this rule if the winning record is set to the
-        ///     default value in the master record.
-        /// </summary>
-        [JsonProperty(PropertyName = "OnlyIfDefault", NullValueHandling = NullValueHandling.Ignore)]
-        public bool OnlyIfDefault { get; set; }
-
-        /// <summary>
-        ///     When a rule is in a group this is the method called to tell the rule the group it
-        ///     belongs to. Will also then validate this rule.
-        /// </summary>
-        /// <returns>True if rule validation passes, else false.</returns>
-        /// <exception cref="Exception">
-        ///     If rule already claimed by a group, or if if filters for types the parent don't include.
-        /// </exception>
-        public bool ClaimAndValidate (GSPGroup group)
-        {
-            if (Group is not null)
-                throw new Exception("Rule already claimed.");
-
-            Group = group;
-
-            if (!Types.Any())
-                Types = Group.Types;  // This may also be NONE but we do extra check for that after validation.
-            else if (Group.Types.Any() && Types.Any(t => !Group.Types.Contains(t)))
-                throw new Exception("Record under group tries to filter for Type(s) that are excluded at by group.");
-
-            if (Priority != 0)
-                Global.Logger.WriteLog(LogLevel.Error, LogType.GeneralConfigFailure, "You have defined a rule priority, for a rule in a group. Group member priorities are ignored. Order in file is processing order.", ClassLogCode, includePrefix: GetLogRuleID());
-
-            bool valid = Validate();
-            if (!Types.Any())
-                Types = Global.Game.AllRecordTypes();
-
-            return valid;
-        }
-
-        #region GetValues
-
-        private readonly Dictionary<FilterOperation, object?> fillCache = [];
-        private readonly Dictionary<FilterOperation, (IEnumerable<ModKey>, string[])?> forwardCache = [];
-        private readonly Dictionary<FilterOperation, object?> matchCache = [];
-
-        /// <summary>
-        ///     Get the value data for a selected rule's action value key parsed to selected class type.
-        /// </summary>
-        public bool TryGetFillValueAs<T> (FilterOperation key, out T? valueAs) => tryGetValueAs(Fill, fillCache, key, out _, out valueAs);
-
-        /// <summary>
-        ///     Gets forward data for current key. If indexed by field, will always have 1 field,
-        ///     but 1 or more mods outputted. If indexed by mod, will always have 1 mod, but 1 or
-        ///     more fields outputted.
-        /// </summary>
-        /// <param name="proKeys">Used to see what order mods should be returned in.</param>
-        /// <param name="key"></param>
-        /// <param name="mods">
-        ///     Output of listed mods. List will be in order entered. If no mods were entered when
-        ///     indexed by field, then will populate with all mods.
-        /// </param>
-        /// <param name="fields">Output of field names from config for current key</param>
-        /// <returns>
-        ///     If valid combination of fields and mods found in config will return true, else false.
-        /// </returns>
-        public bool TryGetForward (ProcessingKeys proKeys, FilterOperation key, [NotNullWhen(true)] out IEnumerable<ModKey>? mods, [NotNullWhen(true)] out string[]? fields)
-        {
-            if (forwardCache.TryGetValue(key, out var value))
-            {
-                if (value is null || !value.HasValue)
-                {
-                    mods = null;
-                    fields = null;
-                    return false;
-                }
-
-                (mods, fields) = value.Value;
-
-                return true;
-            }
-
-            bool sortMods = proKeys.Rule.HasForwardOption(ForwardOptions._sortMods);
-            List<ModKey> buildMods = [];
-            List<string> buildFields = [];
-
-            if (ForwardOptions.HasFlag(ForwardOptions.IndexedByField))
-            {
-                buildFields.Add(key.Value);
-                if (tryGetValueAs(Forward, [], key, out _, out List<ModKeyListOperation>? values))
-                {
-                    bool hasIncludeMods = values?.Any(m => m.Operation == ListLogic.Default) ?? false;
-                    bool hasExcludeMods = values?.Any(m => m.Operation == ListLogic.NOT) ?? false;
-                    sortMods |= hasExcludeMods || !values.SafeAny();
-
-                    if (hasIncludeMods && hasExcludeMods)
-                    {
-                        Global.Logger.WriteLog(LogLevel.Error, LogType.GeneralConfigFailure, "When indexed by field, array of mods must not include both include and excluded mods.", ClassLogCode);
-                    }
-                    else
-                    {
-                        if (hasExcludeMods || !values.SafeAny())
-                        {
-                            // Include all mods except excluded
-                            foreach (var addMod in Global.Game.LoadOrder)
-                            {
-                                if (addMod.Enabled // Mod must be enabled
-                                    && (!hasExcludeMods || !values.SafeAny(v => v.Value.Equals(addMod.ModKey)))) // Exclude anything excluded in config
-                                {
-                                    buildMods.Add(addMod.ModKey);
-                                }
-                            }
-                        }
-                        else
-                        {
-                            // Just include mods listed against field
-                            foreach (var mod in values ?? [])
-                                buildMods.Add(mod.Value);
-                        }
-                    }
-                }
-            }
-            else
-            {
-                if (ModKey.TryFromFileName(new FileName(key.Value), out var mod))
-                    buildMods.Add(mod);
-
-                if (tryGetValueAs(Forward, [], key, out _, out List<string>? values))
-                    buildFields = values ?? buildFields;
-            }
-
-            if (buildMods.Count != 0 && buildFields.Count != 0)
-            {
-                mods = sortMods
-                     ? proKeys.Rule.HasForwardOption(ForwardOptions._randomMod)
-                         ? buildMods.OrderBy(_ => proKeys.GetRandom().Next())
-                         : buildMods.OrderByDescending(Global.Game.LoadOrder.IndexOf) // Descending so same highest to lowest priority as manually entered mod list
-                     : buildMods;
-
-                fields = [.. buildFields];
-
-                forwardCache.Add(key, (mods, fields));
-
-                return true;
-            }
-
-            forwardCache.Add(key, null);
-            mods = null;
-            fields = null;
-            return false;
-        }
-
-        public bool TryGetMatchValueAs<T> (FilterOperation key, out bool fromCache, out T? valueAs) => tryGetValueAs(Match, matchCache, key, out fromCache, out valueAs);
-
-        /// <summary>
-        ///     Get the value data for a selected rule's action value key parsed to selected class type.
-        /// </summary>
-        private static bool tryGetValueAs<T> (Dictionary<FilterOperation, JToken> values, Dictionary<FilterOperation, object?> cache, FilterOperation key, out bool fromCache, out T? valueAs)
-        {
-            if (cache.TryGetValue(key, out object? cachedValue))
-            {
-                fromCache = true;
-                Global.Logger.WriteLog(LogLevel.Trace, LogType.Cache, $"Got value for {key.Value} from cache.", ClassLogCode);
-
-                if (cachedValue is T v)
-                {
-                    valueAs = v;
-                    return true;
-                }
-                else
-                {
-                    valueAs = default;
-
-                    // If value is not null then failed to cast to correct type
-                    return cachedValue is null;
-                }
-            }
-
-            fromCache = false;
-            valueAs = default;
-            bool valid = values.TryGetValue(key, out var jsonValue);
-
-            if (valid && jsonValue is not null)
-                valueAs = jsonValue.Deserialize<T>();
-
-            cache.Add(key, valueAs);
-            return valid;
-        }
-
-        #endregion GetValues
-
-        /// <inheritdoc />
-        public override int GetHashCode ()
-        {
-            if (HashCode == 0)
-            {
-                var hash = new HashCode();
-                hash.Add(base.GetHashCode());
-
-                hash.Add(ForwardOptions);
-                hash.Add(OnlyIfDefault);
-                if (EditorID is not null)
-                    hash.AddEnumerable(EditorID);
-                if (FormID is not null)
-                    hash.AddEnumerable(FormID);
-                if (Match is not null)
-                    hash.AddDictionary(Match);
-                if (Fill is not null)
-                    hash.AddDictionary(Fill);
-                if (Forward is not null)
-                    hash.AddDictionary(Forward);
-
-                HashCode = hash.ToHashCode();
-            }
-
-            return HashCode;
-        }
-
-        public override string GetLogRuleID () => Group is null ? base.GetLogRuleID() : $"{Group.GetLogRuleID()}.{ConfigRule}";
-
-        /// <summary>
-        ///     Checks if rule has a specific Forward Option flag set.
-        /// </summary>
-        /// <param name="flag">Flag you want to check for.</param>
-        /// <returns>True if flag is set on this rule.</returns>
-        public bool HasForwardOption (ForwardOptions flag) => ForwardOptions.HasFlag(flag);
-
-        /// <summary>
-        ///     Checks if rule filter(s) match current context's record.
-        /// </summary>
-        /// <param name="context"></param>
-        /// <param name="Origin"></param>
-        /// <returns></returns>
-        public override bool Matches (ProcessingKeys proKeys)
-        {
-            if (!base.Matches(proKeys))
-                return false;
-
-            if (FormID is not null && !MatchesHelper.Matches(proKeys.Record.FormKey, FormID, Global.Settings.Logging.NoisyLogs.MatchLogs.IncludeFormID))
-                return false;
-
-            if (EditorID is not null && !MatchesHelper.Matches(proKeys.Record.EditorID, EditorID, Global.Settings.Logging.NoisyLogs.MatchLogs.IncludeEditorID))
-                return false;
-
-            foreach (var x in Match)
-            {
-                if (!proKeys.SetProperty(x.Key, x.Key.Value, ClassLogCode))
-                    return false;
-
-                if (!proKeys.Property.Action.CanMatch())
-                {
-                    Global.Logger.WriteLog(LogLevel.Error, LogType.MatchFailure, $"Matched {x.Key}: No match enabled action for field.", ClassLogCode);
-                    return false;
-                }
-
-                Global.Logger.LogAction($"{proKeys.Property.Action.GetType().GetClassName()}.{nameof(IRecordAction.MatchesRule)}", ClassLogCode);
-                if (!proKeys.Property.Action.MatchesRule(proKeys))
-                    return false;
-            }
-
-            Global.Logger.CurrentPropertyName = null; // Clear property name after matching so next rule doesn't use same property name.
-
-            return true;
-        }
-
-        /// <inheritdoc />
-        public override bool Validate ()
-        {
-            if (!base.Validate())
-                return false;
-
-            if (!Types.Any())
-            {
-                Global.Logger.WriteLog(LogLevel.Critical, LogType.GeneralConfigFailure, "Rules must specify record type(s) either directly in the rule or via type(s) added at a group level.", ClassLogCode, includePrefix: GetLogRuleID());
-                return false;
-            }
-
-            return true;
-        }
-    }
-}
