+++ GenericSynthesisPatcher/Helpers/Graph/RecordGraph.cs
@@ -2,7 +2,7 @@ using System.Diagnostics.CodeAnalysis;
 
 using Common;
 
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Microsoft.Extensions.Logging;
