+++ GenericSynthesisPatcher/Helpers/Graph/RecordNodeBase.cs
@@ -2,7 +2,7 @@ using System.Diagnostics.CodeAnalysis;
 
 using DynamicData;
 
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Microsoft.Extensions.Logging;
