+++ GenericSynthesisPatcher/Helpers/MatchesHelper.cs
@@ -1,6 +1,6 @@
 using Common;
 
-using GenericSynthesisPatcher.Games.Universal.Json.Operations;
+using GenericSynthesisPatcher.Rules.Operations;
 
 using Microsoft.Extensions.Logging;
