+++ GenericSynthesisPatcher/Rules/Operations/ModKeyListOperation.cs
@@ -2,7 +2,7 @@ using Mutagen.Bethesda.Plugins;
 
 using Noggog;
 
-namespace GenericSynthesisPatcher.Games.Universal.Json.Operations
+namespace GenericSynthesisPatcher.Rules.Operations
 {
     public class ModKeyListOperation : ListOperationBase<ModKey>
     {
